+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/rule/security/ApexXSSFromURLParamRule.java
@@ -11,6 +11,7 @@ import java.util.Set;
 import net.sourceforge.pmd.lang.apex.ast.ASTAssignmentExpression;
 import net.sourceforge.pmd.lang.apex.ast.ASTBinaryExpression;
 import net.sourceforge.pmd.lang.apex.ast.ASTFieldDeclaration;
+import net.sourceforge.pmd.lang.apex.ast.ASTMethod;
 import net.sourceforge.pmd.lang.apex.ast.ASTMethodCallExpression;
 import net.sourceforge.pmd.lang.apex.ast.ASTReturnStatement;
 import net.sourceforge.pmd.lang.apex.ast.ASTUserClass;
@@ -100,7 +101,10 @@ public class ApexXSSFromURLParamRule extends AbstractApexRule {
 
         ASTMethodCallExpression methodCall = node.getFirstChildOfType(ASTMethodCallExpression.class);
         if (methodCall != null) {
-            processInlineMethodCalls(methodCall, data, true);
+            String retType = getReturnType(node);
+            if ("string".equalsIgnoreCase(retType)) {
+                processInlineMethodCalls(methodCall, data, true);
+            }
         }
 
         List<ASTVariableExpression> nodes = node.findChildrenOfType(ASTVariableExpression.class);
@@ -112,6 +116,15 @@ public class ApexXSSFromURLParamRule extends AbstractApexRule {
         }
 
         return data;
+    }
+
+    private String getReturnType(ASTReturnStatement node) {
+        ASTMethod method = node.getFirstParentOfType(ASTMethod.class);
+        if (method != null) {
+            return method.getNode().getMethodInfo().getReturnType().getApexName();
+        }
+
+        return "";
     }
 
     private boolean isEscapingMethod(ASTMethodCallExpression methodNode) {
@@ -160,8 +173,17 @@ public class ApexXSSFromURLParamRule extends AbstractApexRule {
             if (Helper.isMethodCallChain(right, URL_PARAMETER_METHOD)) {
                 ASTVariableExpression left = node.getFirstChildOfType(ASTVariableExpression.class);
 
+                String varType = null;
+
+                if (node instanceof ASTVariableDeclaration) {
+                    varType = ((ASTVariableDeclaration) node).getNode().getLocalInfo().getType().getApexName();
+
+                }
+
                 if (left != null) {
-                    urlParameterStrings.add(Helper.getFQVariableName(left));
+                    if (varType == null || !"id".equalsIgnoreCase(varType)) {
+                        urlParameterStrings.add(Helper.getFQVariableName(left));
+                    }
                 }
             }
 
@@ -190,7 +212,15 @@ public class ApexXSSFromURLParamRule extends AbstractApexRule {
     private void processVariableAssignments(AbstractApexNode<?> node, Object data, final boolean reverseOrder) {
         ASTMethodCallExpression methodCallAssignment = node.getFirstChildOfType(ASTMethodCallExpression.class);
         if (methodCallAssignment != null) {
-            processInlineMethodCalls(methodCallAssignment, data, false);
+
+            String varType = null;
+            if (node instanceof ASTVariableDeclaration) {
+                varType = ((ASTVariableDeclaration) node).getNode().getLocalInfo().getType().getApexName();
+            }
+
+            if (varType == null || !"id".equalsIgnoreCase(varType)) {
+                processInlineMethodCalls(methodCallAssignment, data, false);
+            }
         }
 
         List<ASTVariableExpression> nodes = node.findChildrenOfType(ASTVariableExpression.class);