+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/rule/security/Helper.java
@@ -181,6 +181,13 @@ public final class Helper {
         return sb.toString();
     }
 
+    static String getVariableType(final ASTField variable) {
+        Field n = variable.getNode();
+        StringBuilder sb = new StringBuilder().append(n.getDefiningType().getApexName()).append(":")
+                .append(n.getFieldInfo().getName());
+        return sb.toString();
+    }
+    
     static String getFQVariableName(final ASTFieldDeclaration variable) {
         FieldDeclaration n = variable.getNode();
         String name = "";